migrated tests to a test.py file. created script to fetch specific blog post.
